refactor analytics to use supabase
